pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conductor-sdk"
version = "0.1.0"
description = "Tiny SDK to read request data and publish analysis outputs/results.json for Conductor."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "you@example.com"}]
dependencies = ["google-cloud-storage>=2.17.0"]

[project.optional-dependencies]
analysis = ["pandas>=2.1", "numpy>=1.26", "matplotlib>=3.8"]

[tool.setuptools]
packages = ["conductor_sdk"]
